# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Spotify Flask API",
    "description": "API documentation for Spotify integration with Flask",
    "version": "1.0.1"
  },
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "tags": [
    {
      "name": "Authentication and Token Management",
      "description": "Endpoints related to user authentication and token management."
    },
    {
      "name": "User Information",
      "description": "Endpoints for retrieving user-specific information from Spotify."
    },
    {
      "name": "User Media",
      "description": "Endpoints for retrieving user media information from Spotify."
    }
  ],
  "paths": {
    "/api/Login": {
      "get": {
        "summary": "Get Spotify access token",
        "tags": [
          "Authentication and Token Management"
        ],
        "description": "Retrieve an access token for authenticating with Spotify's API.",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "access_token": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Something went wrong on the server.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Logout": {
      "get": {
        "summary": "Log out the user",
        "tags": [
          "Authentication and Token Management"
        ],
        "description": "Log out the currently authenticated user.",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Something went wrong on the server.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/MyProfile": {
      "get": {
        "summary": "Get user profile from Spotify",
        "tags": [
          "User Information"
        ],
        "description": "Retrieve the Spotify profile information for the authenticated user.",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object"
            }
          },
          "401": {
            "description": "Unauthorized. Authentication required.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Something went wrong on the server.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/MyPicture": {
      "get": {
        "summary": "Get user pictures from Spotify",
        "tags": [
          "User Information"
        ],
        "description": "Retrieve the Spotify profile pictures for the authenticated user.",
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Authentication required.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Not Found. The specified resource was not found.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Something went wrong on the server.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/FavArtists": {
      "get": {
        "summary": "Get user favorite artists from Spotify",
        "tags": [
          "User Media"
        ],
        "description": "Retrieve the favorite artists for the authenticated user from Spotify.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of artists to return",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "time_range",
            "in": "query",
            "description": "Time range to return (e.g., short_term, medium_term, long_term)",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Authentication required.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Invalid request parameters.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Something went wrong on the server.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/FavSongs": {
      "get": {
        "summary": "Get user favorite songs from Spotify",
        "tags": [
          "User Media"
        ],
        "description": "Retrieve the favorite songs for the authenticated user from Spotify.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of songs to return",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "time_range",
            "in": "query",
            "description": "Time range to return (e.g., short_term, medium_term, long_term)",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "artist": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Authentication required.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Invalid request parameters.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Something went wrong on the server.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Recommendations": {
      "get": {
        "summary": "Get song recommendations from Spotify",
        "tags": [
          "User Media"
        ],
        "description": "Retrieve song recommendations for the authenticated user from Spotify.",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "description": "Number of recommendations to return",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "seed_artists",
            "in": "query",
            "description": "Comma-separated list of seed artist IDs",
            "required": false,
            "type": "string"
          },
          {
            "name": "seed_tracks",
            "in": "query",
            "description": "Comma-separated list of seed track IDs",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "artist": {
                    "type": "string"
                  },
                  "uri": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized. Authentication required.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Invalid request parameters.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Something went wrong on the server.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/ReceiveToken": {
      "put": {
        "summary": "Receive token",
        "tags": [
          "Authentication and Token Management"
        ],
        "description": "Receive and store the Spotify access token.",
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "description": "Token to store in the database",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Token received successfully",
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Invalid request parameters.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error. Something went wrong on the server.",
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/add_to_playlist": {
  "post": {
    "summary": "Add songs to a new playlist",
    "tags": [
      "User Media"
    ],
    "description": "Create a new playlist and add provided songs to it.",
    "parameters": [
      {
        "name": "tracks",
        "in": "body",
        "description": "List of track URIs to add to the playlist",
        "required": true,
        "schema": {
          "type": "object",
          "properties": {
            "tracks": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "uri": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    ],
    "responses": {
      "200": {
        "description": "Songs added successfully",
        "schema": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Songs added successfully!"
            }
          }
        }
      },
      "400": {
        "description": "Bad Request. Invalid request parameters.",
        "schema": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "No tracks provided"
            }
          }
        }
      },
      "500": {
        "description": "Internal Server Error. Something went wrong on the server.",
        "schema": {
          "type": "object",
          "properties": {
            "message": {
              "type": "string",
              "example": "Failed to create playlist"
            }
          }
        }
      }
    }
  }
}
  }
}